+++ Implementations/Sponge/src/main/java/cz/neumimto/rpg/sponge/effects/SingleResistanceValueEffect.java
@@ -26,14 +26,14 @@ public abstract class SingleResistanceValueEffect extends EffectBase<Float> {
     @Override
     public void onApply(IEffect self) {
         super.onApply(self);
-        float characterProperty = getConsumer().getProperty(getPropertyId());
+        double characterProperty = getConsumer().getProperty(getPropertyId());
         getConsumer().setProperty(getPropertyId(), characterProperty + getValue());
     }
 
     @Override
     public void onRemove(IEffect self) {
         super.onRemove(self);
-        float characterProperty = getConsumer().getProperty(getPropertyId());
+        double characterProperty = getConsumer().getProperty(getPropertyId());
         getConsumer().setProperty(getPropertyId(), characterProperty - getValue());
     }
 }
